# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Quick Menu API
  version: "1.0.0"
  description: |
    Quick Menu backend API — QR-based menu, real-time orders (WebSocket), admin dashboards.
servers:
  - url: http://localhost:8080
    description: Local dev server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        timezone:
          type: string
        currency:
          type: string
        address:
          type: string
    Table:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        qrUrl:
          type: string
          format: uri
...
  parameters:
    restaurantId:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
    tableId:
      name: tableId
      in: path
      required: true
      schema:
        type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/restaurants:
    post:
      summary: Create a restaurant (admin)
      tags: [Restaurants]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
    get:
      summary: List restaurants (paginated)
      tags: [Restaurants]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            default: id,asc
            description: "sort=property,asc|desc (repeatable)"
      responses:
        '200':
          description: Page of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  pageable:
                    type: object
                  totalElements:
                    type: integer
  /api/restaurants/{restaurantId}:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    get:
      summary: Get restaurant details
      tags: [Restaurants]
      responses:
        '200':
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Not found
    patch:
      summary: Update restaurant
      tags: [Restaurants]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
    delete:
      summary: Delete restaurant
      tags: [Restaurants]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  # Tables (QR)
  /api/restaurants/{restaurantId}/tables:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
    post:
      summary: Create a table (generates QR url)
      tags: [Tables]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
    get:
      summary: List tables (paginated)
      tags: [Tables]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: sort
          in: query
          schema:
            type: string
            default: id,asc
      responses:
        '200':
          description: Page of tables
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Table'
                  totalElements:
                    type: integer
  /api/restaurants/{restaurantId}/tables/{tableId}:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
      - $ref: '#/components/parameters/tableId'
    get:
      summary: Get table
      tags: [Tables]
      responses:
        '200':
          description: Table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
    delete:
      summary: Delete table
      tags: [Tables]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /api/restaurants/{restaurantId}/tables/{tableId}/qr.png:
    parameters:
      - $ref: '#/components/parameters/restaurantId'
      - $ref: '#/components/parameters/tableId'
    get:
      summary: Get table QR PNG (image/png)
      tags: [Tables]
      responses:
        '200':
          description: PNG image of QR (binary)
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Not found

  # Bell (customer pings staff)
...